feat: signin, signup

+++ src/migrations/20230522141013_user.js
@@ -1,18 +1,22 @@
 export async function up(knex) {
   knex.schema.hasTable('users').then((exists) => {
     if (!exists) {
-      return knex.schema
-        .createTable('users', function (table) {
-          table.increments('id');
-          table.string('username', 255).notNullable();
-          table.string('type', 255).notNullable();
-          table.string('password', 255).notNullable();
-        });
+      return knex.schema.createTable('users', (t) => {
+        t.increments('id');
+        t.string('username', 255).notNullable();
+        t.string('type', 255).notNullable();
+        t.string('password', 255).notNullable();
+        t.date('created_at', new Date().toISOString()).notNullable();
+        t.date('updated_at', new Date().toISOString()).notNullable();
+      });
     }
   });
 }
 
-
 export async function down(knex) {
-  return knex.schema.dropTable('users');
+  knex.schema.hasTable('users').then((exists) => {
+    if (exists) {
+      return knex.schema.dropTable('users');
+    }
+  });
 }
